Create README.md
Was never into computers as a kid or a teenager and now I find myself trying to catch up and learn as much as possible. I see things and have my whole life in a different way. I look at something and say to myself how can this be done more efficiently. I currently own a company that forced me to go down this path and I love it. I own an automotive calibrations business. In short we set up a stand with a target and configure the lane camera or radar. This has led to diagnostics and programming which led to learning how the programs are written and need to understand how things work. I like the engineering part of things more than html coding. I am excited to learn and stoked there is a place online like GitHub and I hope to learn and contribute as much as possible. Thanks everyone!